Use ETag for conditional requests. Fixes #2.
The URIs and ETags of all GitHub API responses that are used to construct the feed are hashed together to make the ETag for the feed itself.

This doesn't save any processing time on the server, but benefits the client by allowing it to easily reuse its own cached data.