Use CancellationToken to avoid unnecessary steps.
Previously, most methods would check if m_result had been set yet and, if so, do nothing. Now, we pass a CancellationToken to ContinueWith, and cancel that token when the final ActionResult is known. This avoids calling unnecessary methods and lets us skip straight to returning that ActionResult to the client.